+++ src/lib/st-textarea/index.ts
@@ -1,0 +1,2 @@
+export * from './st-textarea.module';
+export * from './st-textarea.error.model';