+++ src/lib/st-textarea/st-textarea.error.model.ts
@@ -1,0 +1,11 @@
+import { TranslateableElement } from '../utils/egeo-resolver/egeo-resolve-model';
+
+export interface StTextareaError {
+  generic?: string;
+  required?: string;
+}
+
+export interface StTextareaErrorSchema {
+  generic?: TranslateableElement;
+  required?: TranslateableElement;
+}